moved test file to where admin connection is possible
